fix : 마이페이지 들어갈 수 있게 설정

+++ src/main/java/live/ioteatime/apiservice/domain/Organization.java
@@ -15,7 +15,7 @@ public class Organization {
     private int id;
     private String name;
     @Column(name = "electricity_budget")
-    private BigInteger electricityBudget;
+    private Long electricityBudget;
     @Column(name = "organization_code")
     private String organizationCode;
 }
